Move functions from helper classes
