+++ Kitodo/src/main/java/de/sub/goobi/export/dms/AutomaticDmsExportWithoutHibernate.java
@@ -14,7 +14,6 @@ package de.sub.goobi.export.dms;
 import de.sub.goobi.config.ConfigCore;
 import de.sub.goobi.config.ConfigProjects;
 import de.sub.goobi.export.download.ExportMetsWithoutHibernate;
-import de.sub.goobi.helper.FilesystemHelper;
 import de.sub.goobi.helper.Helper;
 import de.sub.goobi.helper.tasks.EmptyTask;
 import de.sub.goobi.metadaten.MetadatenVerifizierungWithoutHibernate;
@@ -377,7 +376,8 @@ public class AutomaticDmsExportWithoutHibernate extends ExportMetsWithoutHiberna
                  */
                 User myUser = (User) Helper.getManagedBeanValue("#{LoginForm.myBenutzer}");
                 try {
-                    FilesystemHelper.createDirectoryForUser(zielTif.getAbsolutePath(), myUser.getLogin());
+                    serviceManager.getFileService().createDirectoryForUser(zielTif.getAbsolutePath(),
+                            myUser.getLogin());
                 } catch (Exception e) {
                     if (task != null) {
                         task.setException(e);
