+++ Kitodo/src/main/java/de/sub/goobi/export/download/ExportMets.java
@@ -16,7 +16,6 @@ import de.sub.goobi.config.ConfigProjects;
 import de.sub.goobi.export.dms.ExportDms;
 import de.sub.goobi.export.dms.ExportDms_CorrectRusdml;
 import de.sub.goobi.forms.LoginForm;
-import de.sub.goobi.helper.FilesystemHelper;
 import de.sub.goobi.helper.Helper;
 import de.sub.goobi.helper.VariableReplacer;
 import de.sub.goobi.helper.exceptions.ExportFileException;
@@ -137,7 +136,7 @@ public class ExportMets {
         User myBenutzer = (User) Helper.getManagedBeanValue("#{LoginForm.myBenutzer}");
         if (myBenutzer != null) {
             try {
-                FilesystemHelper.createDirectoryForUser(target, myBenutzer.getLogin());
+                serviceManager.getFileService().createDirectoryForUser(target, myBenutzer.getLogin());
             } catch (Exception e) {
                 Helper.setFehlerMeldung("Export canceled, could not create destination directory: " + inTargetFolder,
                         e);
