+++ Kitodo/src/main/java/org/goobi/io/BackupFileRotation.java
@@ -11,14 +11,13 @@
 
 package org.goobi.io;
 
-import de.sub.goobi.helper.FilesystemHelper;
-
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FilenameFilter;
 import java.io.IOException;
 
 import org.apache.log4j.Logger;
+import org.kitodo.services.ServiceManager;
 
 /**
  * Creates backup for files in a given directory that match a regular
@@ -45,6 +44,8 @@ public class BackupFileRotation {
     private int numberOfBackups;
     private String format;
     private String processDataDirectory;
+
+    private ServiceManager serviceManager = new ServiceManager();
 
     /**
      * Start the configured backup.
@@ -115,7 +116,7 @@ public class BackupFileRotation {
         rotateBackupFilesFor(fileName);
 
         String newName = fileName + ".1";
-        FilesystemHelper.renameFile(fileName, newName);
+        serviceManager.getFileService().renameFile(fileName, newName);
     }
 
     private void rotateBackupFilesFor(String fileName) throws IOException {
@@ -130,7 +131,7 @@ public class BackupFileRotation {
             String oldName = fileName + "." + (count - 1);
             String newName = fileName + "." + count;
             try {
-                FilesystemHelper.renameFile(oldName, newName);
+                serviceManager.getFileService().renameFile(oldName, newName);
             } catch (FileNotFoundException oldNameNotYetPresent) {
                 if (myLogger.isDebugEnabled()) {
                     myLogger.debug(oldName + " does not yet exist >>> nothing to do");
